02-DB-Fastapi/pyproject.toml
[project]
name = "02-db-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
